기본 기능 완성

+++ src/main/java/org/jeskey/service/BoardService.java
@@ -9,11 +9,11 @@ public interface BoardService {
 	
 	public List<BoardDTO> getList(PageDTO dto);
 	
-	public int insert(BoardDTO dto);
+	public Long insert(BoardDTO dto);
 	
 	public BoardDTO getOne(Long bno);
 	
-	public int update(BoardDTO dto);
+	public Long update(BoardDTO dto);
 	
 	public int delete(Long bno);
